Update modman
The problem here is that if you leave it as it is, and after you have installed the latest patch on your magento installation, since the composer strategy is changed to copy, the mapping is completely wrong. So every time I do a composer install, I see Adyen/Payment/Fee/Fee. Then again  Adyen/Payment/Fee/Fee/Fee and on and on. This needs to fixed and released because can cause major problems.