Adyen Payment PHP 7 compatibility
In the function "addMethodsToConfig" the first parameter is expected to be a Varien_Event_Observer. In PHP 5 it is possible to pass null to this parameter. In PHP 7 this is not possible anymore.

Solution to this issue, Pass an empty Varien_Event_Observer as parameter since no where in the function is checked on null, no behavior will change.